requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
python-dotenv==1.0.0
seam==1.71.0
requests==2.31.0
email-validator==2.0.0
gunicorn==21.2.0
pytest==7.3.1
pytest-flask==1.2.0
pytz==2024.1 